Share the model picker between the composer and the thread header
Move the composer's model chip into StudioModelPicker so the Converse
thread header offers the same filtered, installed-first menu. Parameter
counts in ids read the way model cards print them ("4b" -> "4B").
